requirements.txt
tensorflow==2.15.0
pillow==10.2.0